back/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2023.7.22
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.5
geographiclib==2.0
geopy==2.4.1
httpx>=0.27
Interface==2.11.1
joblib==1.4.2
numpy<2
pandas==2.2.3
prometheus_client==0.21.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.36
starlette==0.41.3
threadpoolctl==3.5.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
zope.event==5.0
zope.interface==6.0
zope.schema==7.0.1

# --- Added in Phase 0: Causal Decision Copilot ---
# Causal inference
dowhy==0.11.1
econml==0.15.1
networkx<3.3  # dowhy 0.11 uses nx.algorithms.d_separated, removed in 3.3
# Optimization
ortools==9.11.4210
cvxpy==1.5.3
pulp==2.9.0
# Simulation
simpy==4.1.1
# Explainability
shap<0.44.0
# Experiment tracking & patent audit trail
mlflow==2.17.2
# Real-time infra (already referenced by main.py)
fastapi-limiter==0.1.6
redis==5.2.0
slack-sdk==3.33.3
# Time-series forecasting (referenced by main.py: ARIMA)
statsmodels==0.14.4
# Migrations
alembic>=1.13
# Pin networkx because DoWhy 0.11 still imports nx.algorithms.d_separated
# which was removed in networkx 3.3.
networkx<3.3
# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
# (httpx already listed above — used by TestClient)
